Forwardport old restore deprecation
Forwardport 446b48f3528718c287823fc4520e04dc9117cb5e from 1.25
