+++ state/backups/backups_linux.go
@@ -7,15 +7,49 @@ package backups
 
 import (
 	"fmt"
+	"strconv"
 
 	"github.com/juju/errors"
 	"github.com/juju/names"
+	"github.com/juju/utils/shell"
 
 	"github.com/juju/juju/agent"
 	"github.com/juju/juju/juju/paths"
+	"github.com/juju/juju/mongo"
 	"github.com/juju/juju/network"
+	"github.com/juju/juju/service"
 	"github.com/juju/juju/state"
 )
+
+func ensureMongoService(agentConfig agent.Config) error {
+	var oplogSize int
+	if oplogSizeString := agentConfig.Value(agent.MongoOplogSize); oplogSizeString != "" {
+		var err error
+		if oplogSize, err = strconv.Atoi(oplogSizeString); err != nil {
+			return errors.Annotatef(err, "invalid oplog size: %q", oplogSizeString)
+		}
+	}
+
+	var numaCtlPolicy bool
+	if numaCtlString := agentConfig.Value(agent.NumaCtlPreference); numaCtlString != "" {
+		var err error
+		if numaCtlPolicy, err = strconv.ParseBool(numaCtlString); err != nil {
+			return errors.Annotatef(err, "invalid numactl preference: %q", numaCtlString)
+		}
+	}
+
+	si, ok := agentConfig.StateServingInfo()
+	if !ok {
+		return errors.Errorf("agent config has no state serving info")
+	}
+
+	err := mongo.EnsureServiceInstalled(agentConfig.DataDir(),
+		agentConfig.Value(agent.Namespace),
+		si.StatePort,
+		oplogSize,
+		numaCtlPolicy)
+	return errors.Annotate(err, "cannot ensure that mongo service start/stop scripts are in place")
+}
 
 // Restore handles either returning or creating a state server to a backed up status:
 // * extracts the content of the given backup file and:
@@ -24,17 +58,17 @@ import (
 // * updates existing db entries to make sure they hold no references to
 // old instances
 // * updates config in all agents.
-func (b *backups) Restore(backupId string, args RestoreArgs) error {
+func (b *backups) Restore(backupId string, args RestoreArgs) (names.Tag, error) {
 	meta, backupReader, err := b.Get(backupId)
 	if err != nil {
-		return errors.Annotatef(err, "could not fetch backup %q", backupId)
+		return nil, errors.Annotatef(err, "could not fetch backup %q", backupId)
 	}
 
 	defer backupReader.Close()
 
 	workspace, err := NewArchiveWorkspaceReader(backupReader)
 	if err != nil {
-		return errors.Annotate(err, "cannot unpack backup file")
+		return nil, errors.Annotate(err, "cannot unpack backup file")
 	}
 	defer workspace.Close()
 
@@ -42,86 +76,115 @@ func (b *backups) Restore(backupId string, args RestoreArgs) error {
 	version := meta.Origin.Version
 	backupMachine := names.NewMachineTag(meta.Origin.Machine)
 
+	if err := mongo.StopService(agent.Namespace); err != nil {
+		return nil, errors.Annotate(err, "cannot stop mongo to replace files")
+	}
+
 	// delete all the files to be replaced
 	if err := PrepareMachineForRestore(); err != nil {
-		return errors.Annotate(err, "cannot delete existing files")
+		return nil, errors.Annotate(err, "cannot delete existing files")
 	}
+	logger.Infof("deleted old files to place new")
 
 	if err := workspace.UnpackFilesBundle(filesystemRoot()); err != nil {
-		return errors.Annotate(err, "cannot obtain system files from backup")
+		return nil, errors.Annotate(err, "cannot obtain system files from backup")
 	}
-
-	if err := updateBackupMachineTag(backupMachine, args.NewInstTag); err != nil {
-		return errors.Annotate(err, "cannot update paths to reflect current machine id")
-	}
+	logger.Infof("placed new files")
 
 	var agentConfig agent.ConfigSetterWriter
 	// The path for the config file might change if the tag changed
 	// and also the rest of the path, so we assume as little as possible.
 	datadir, err := paths.DataDir(args.NewInstSeries)
 	if err != nil {
-		return errors.Annotate(err, "cannot determine DataDir for the restored machine")
+		return nil, errors.Annotate(err, "cannot determine DataDir for the restored machine")
 	}
-	agentConfigFile := agent.ConfigPath(datadir, args.NewInstTag)
+	agentConfigFile := agent.ConfigPath(datadir, backupMachine)
 	if agentConfig, err = agent.ReadConfig(agentConfigFile); err != nil {
-		return errors.Annotate(err, "cannot load agent config from disk")
+		return nil, errors.Annotate(err, "cannot load agent config from disk")
 	}
 	ssi, ok := agentConfig.StateServingInfo()
 	if !ok {
-		return errors.Errorf("cannot determine state serving info")
+		return nil, errors.Errorf("cannot determine state serving info")
 	}
-	// The machine tag might have changed, we update it.
-	agentConfig.SetValue("tag", args.NewInstTag.String())
-	apiHostPorts := [][]network.HostPort{
-		network.NewHostPorts(ssi.APIPort, args.PrivateAddress),
-	}
-	agentConfig.SetAPIHostPorts(apiHostPorts)
+	APIHostPorts := network.NewHostPorts(ssi.APIPort, args.PrivateAddress)
+	agentConfig.SetAPIHostPorts([][]network.HostPort{APIHostPorts})
 	if err := agentConfig.Write(); err != nil {
-		return errors.Annotate(err, "cannot write new agent configuration")
+		return nil, errors.Annotate(err, "cannot write new agent configuration")
+	}
+	logger.Infof("wrote new agent config")
+
+	if backupMachine.Id() != "0" {
+		logger.Infof("extra work needed backup belongs to %q machine", backupMachine.String())
+		serviceName := "jujud-" + agentConfig.Tag().String()
+		aInfo := service.NewMachineAgentInfo(
+			agentConfig.Tag().Id(),
+			dataDir,
+			paths.MustSucceed(paths.LogDir(args.NewInstSeries)),
+		)
+
+		// TODO(perrito666) renderer should have a RendererForSeries, for the moment
+		// restore only works on linuxes.
+		renderer, _ := shell.NewRenderer("bash")
+		serviceAgentConf := service.AgentConf(aInfo, renderer)
+		svc, err := service.NewService(serviceName, serviceAgentConf, args.NewInstSeries)
+		if err != nil {
+			return nil, errors.Annotate(err, "cannot generate service for the restored agent.")
+		}
+		if err := svc.Install(); err != nil {
+			return nil, errors.Annotate(err, "cannot install service for the restored agent.")
+		}
+		logger.Infof("new machine service")
 	}
 
+	logger.Infof("mongo service will be reinstalled to ensure its presence")
+	if err := ensureMongoService(agentConfig); err != nil {
+		return nil, errors.Annotate(err, "failed to reinstall service for juju-db")
+	}
+
+	logger.Infof("new mongo will be restored")
 	// Restore mongodb from backup
-	if err := placeNewMongo(workspace.DBDumpDir, version); err != nil {
-		return errors.Annotate(err, "error restoring state from backup")
+	if err := placeNewMongoService(workspace.DBDumpDir, version); err != nil {
+		return nil, errors.Annotate(err, "error restoring state from backup")
 	}
 
 	// Re-start replicaset with the new value for server address
 	dialInfo, err := newDialInfo(args.PrivateAddress, agentConfig)
 	if err != nil {
-		return errors.Annotate(err, "cannot produce dial information")
+		return nil, errors.Annotate(err, "cannot produce dial information")
 	}
 
+	logger.Infof("restarting replicaset")
 	memberHostPort := fmt.Sprintf("%s:%d", args.PrivateAddress, ssi.StatePort)
 	err = resetReplicaSet(dialInfo, memberHostPort)
 	if err != nil {
-		return errors.Annotate(err, "cannot reset replicaSet")
+		return nil, errors.Annotate(err, "cannot reset replicaSet")
 	}
 
 	err = updateMongoEntries(args.NewInstId, args.NewInstTag.Id(), backupMachine.Id(), dialInfo)
 	if err != nil {
-		return errors.Annotate(err, "cannot update mongo entries")
+		return nil, errors.Annotate(err, "cannot update mongo entries")
 	}
 
 	// From here we work with the restored state server
 	mgoInfo, ok := agentConfig.MongoInfo()
 	if !ok {
-		return errors.Errorf("cannot retrieve info to connect to mongo")
+		return nil, errors.Errorf("cannot retrieve info to connect to mongo")
 	}
 
 	st, err := newStateConnection(agentConfig.Environment(), mgoInfo)
 	if err != nil {
-		return errors.Trace(err)
+		return nil, errors.Trace(err)
 	}
 	defer st.Close()
 
-	machine, err := st.Machine(args.NewInstTag.Id())
+	machine, err := st.Machine(backupMachine.Id())
 	if err != nil {
-		return errors.Trace(err)
+		return nil, errors.Trace(err)
 	}
 
 	err = updateMachineAddresses(machine, args.PrivateAddress, args.PublicAddress)
 	if err != nil {
-		return errors.Annotate(err, "cannot update api server machine addresses")
+		return nil, errors.Annotate(err, "cannot update api server machine addresses")
 	}
 
 	// update all agents known to the new state server.
@@ -130,21 +193,20 @@ func (b *backups) Restore(backupId string, args RestoreArgs) error {
 	// agent update failures
 	machines, err := st.AllMachines()
 	if err != nil {
-		return errors.Trace(err)
+		return nil, errors.Trace(err)
 	}
 	if err = updateAllMachines(args.PrivateAddress, machines); err != nil {
-		return errors.Annotate(err, "cannot update agents")
+		return nil, errors.Annotate(err, "cannot update agents")
 	}
 
 	info, err := st.RestoreInfoSetter()
-
 	if err != nil {
-		return errors.Trace(err)
+		return nil, errors.Trace(err)
 	}
 
 	// Mark restoreInfo as Finished so upon restart of the apiserver
 	// the client can reconnect and determine if we where succesful.
 	err = info.SetStatus(state.RestoreFinished)
 
-	return errors.Annotate(err, "failed to set status to finished")
+	return backupMachine, errors.Annotate(err, "failed to set status to finished")
 }
