+++ state/backups/files.go
@@ -6,6 +6,7 @@ package backups
 import (
 	"fmt"
 	"os"
+	"path"
 	"path/filepath"
 	"sort"
 
@@ -130,6 +131,7 @@ func replaceableFoldersFunc() (map[string]os.FileMode, error) {
 		filepath.Join(dataDir, "db"),
 		dataDir,
 		logsDir,
+		path.Join(logsDir, "all-machines.log"),
 	} {
 		dirStat, err := os.Stat(replaceable)
 		if err != nil {
@@ -164,6 +166,9 @@ func PrepareMachineForRestore() error {
 		_, err := os.Stat(toBeRecreated)
 		if err != nil && !os.IsNotExist(err) {
 			return errors.Trace(err)
+		}
+		if !fmode.IsDir() {
+			continue
 		}
 		if err := os.RemoveAll(toBeRecreated); err != nil {
 			return errors.Trace(err)
