+++ state/backups/testing/fakes.go
@@ -7,6 +7,7 @@ import (
 	"io"
 
 	"github.com/juju/errors"
+	"github.com/juju/names"
 	jc "github.com/juju/testing/checkers"
 	"github.com/juju/utils/filestorage"
 	gc "gopkg.in/check.v1"
@@ -96,11 +97,11 @@ func (b *FakeBackups) Remove(id string) error {
 }
 
 // Restore restores a machine to a backed up status.
-func (b *FakeBackups) Restore(bkpId string, args backups.RestoreArgs) error {
+func (b *FakeBackups) Restore(bkpId string, args backups.RestoreArgs) (names.Tag, error) {
 	b.Calls = append(b.Calls, "Restore")
 	b.PrivateAddr = args.PrivateAddress
 	b.InstanceId = args.NewInstId
-	return errors.Trace(b.Error)
+	return nil, errors.Trace(b.Error)
 }
 
 // TODO(ericsnow) FakeStorage should probably move over to the utils repo.
